+++ internal/sudo/sudo.go
@@ -5,9 +5,15 @@ import (
 	"os/exec"
 )
 
-func RunCommand(nitro, machine, command string) error {
+func RunCommand(nitro, machine string, commands ...string) error {
 	// TODO update this for windows support
-	c := exec.Command("sudo", nitro, "-m", machine, command)
+
+	b := []string{nitro, "-m", machine}
+	for _, command := range commands {
+		b = append(b, command)
+	}
+
+	c := exec.Command("sudo", b...)
 
 	c.Stdout = os.Stdout
 	c.Stderr = os.Stderr