+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
  
  ### Changes
+ - `remove` will not prompt for sites to remove if it cannot find any sites in the config file.
+ - `remove` will now remove the machines sites from the hosts file. 
  - `apply` will now create any new databases that it finds in the config file.
  - `destroy` command is now always permanent and the `--permanent` flag has been removed. `destroy` is no longer nested under the machine command.
  - Fixed an error where a user could select MySQL version 5.8.